Create TEST 9
This is the 9th feasibility test for the HAD 2.0 model. The test was successful and tested against the objection if all of the radioisotopes had escaped would this lead to all of the argon escaping? No, it would not. This was tested against the continual plausibility of the model to see if the model would still work in light of accelerated nuclear decay and the possibility of scaping argon.